# backend/BookshelfAPI/DTOs/BookDTO.cs
namespace AuthorsWebAPI.DTOs
{
    public class BookDTO
    {
        public int Id { get; set; }
        public string Title { get; set; }
        public string ImageUrl { get; set; }
        public DateTime PublishedDate { get; set; }
    }
}